Web_Frameworks/Django/Airbnb/pyproject.toml
[project]
name = "airbnb"
version = "0.1.0"
description = ""
authors = [
    {name = "kikiru328",email = "kikiru328@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4"
dependencies = [
    "pillow (>=11.1.0,<12.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "django-environ (>=0.12.0,<0.13.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
